# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateRoleRequest:
    properties:
      name:
        example: Admin
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  dto.PaginatedRoleResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.RoleResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.RoleResponse:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      id:
        example: a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d
        type: string
      name:
        example: Admin
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
    type: object
  dto.UpdatePermissionsRequest:
    properties:
      permission_ids:
        example:
        - a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d
        - f9e8d7c6-b5a4-3210-fedc-ba0987654321
        items:
          type: string
        minItems: 1
        type: array
    required:
    - permission_ids
    type: object
...
  dto.UpdateRoleRequest:
    properties:
      name:
        example: Admin Updated
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  response.ApiResponse:
    properties:
      data: {}
      error:
        example: Error description
        type: string
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: Production-ready API boilerplate built with Gin, GORM, and Redis.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Gin Boilerplate API
  version: "1.0"
paths:
...
  /api/v1/roles:
    get:
      description: Get paginated list of roles with optional filters and sorting.
        Requires 'view_role' permission.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Sorting criteria (e.g. name:asc, created_at:desc or -created_at)
        in: query
        name: sort
        type: string
      - description: Filter by role name (partial match)
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Roles retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginatedRoleResponse'
              type: object
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "403":
          description: Forbidden - Requires view_role permission
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid filter or sorting parameter
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create a new unique role. Requires 'create_role' permission.
      parameters:
      - description: Role creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role created successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "403":
          description: Forbidden - Requires create_role permission
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "409":
          description: Conflict - Role already exists
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid request payload
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create a new role
      tags:
      - Roles
  /api/v1/roles/{id}:
    delete:
      description: Delete a role by UUID. Requires 'delete_role' permission.
      parameters:
      - description: Role UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role deleted successfully
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "403":
          description: Forbidden - Requires delete_role permission
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found - Role not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid UUID format
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - Roles
    get:
      description: Retrieve detailed information of a role including assigned permissions
        by UUID. Requires 'view_role' permission.
      parameters:
      - description: Role UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "403":
          description: Forbidden - Requires view_role permission
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found - Role not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid UUID format
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get role by ID
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Update role information by UUID. Requires 'update_role' permission.
      parameters:
      - description: Role UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Role update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "403":
          description: Forbidden - Requires update_role permission
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found - Role not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid UUID format or request payload
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - Roles
  /api/v1/roles/{id}/permissions:
    delete:
      consumes:
      - application/json
      description: Remove one or more permissions from a role by permission UUIDs.
        Requires 'remove_role_permission' permission.
      parameters:
      - description: Role UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Permission UUIDs to remove
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Permissions removed from role successfully
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "403":
          description: Forbidden - Requires remove_role_permission permission
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found - Role not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid UUID format or request payload
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Remove permissions from role
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Assign one or more permissions to a role by permission UUIDs. Requires
        'add_role_permission' permission.
      parameters:
      - description: Role UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Permission UUIDs to assign
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePermissionsRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Permissions added to role successfully
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "403":
          description: Forbidden - Requires add_role_permission permission
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found - Role not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid UUID format or request payload
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Add permissions to role
      tags:
      - Roles
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token (e.g. "Bearer eyJhbGci...").
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"